backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
SpeechRecognition
google-generativeai
pyttsx3
deepface
opencv-python
pillow
numpy